+++ fintech/src/test/java/ru/kslacker/fintech/service/WeatherServiceImplTest.java
@@ -32,9 +32,9 @@ public class WeatherServiceImplTest {
         UUID region1 = UUID.randomUUID();
         UUID region2 = UUID.randomUUID();
         List<Weather> weather = List.of(
-                new Weather(region1, "1", 1, LocalDateTime.now()),
-                new Weather(region2, "2", -2, LocalDateTime.now()),
-                new Weather(region2, "2", -4, LocalDateTime.now())
+                new Weather(region1, "1", 1.0, LocalDateTime.now()),
+                new Weather(region2, "2", -2.0, LocalDateTime.now()),
+                new Weather(region2, "2", -4.0, LocalDateTime.now())
         );
         when(weatherRepository.findAll()).thenReturn(weather);
 
@@ -53,13 +53,13 @@ public class WeatherServiceImplTest {
         UUID region1 = UUID.randomUUID();
         UUID region2 = UUID.randomUUID();
         List<Weather> weather = List.of(
-                new Weather(region1, "1", 1, LocalDateTime.now()),
-                new Weather(region2, "2", -2, LocalDateTime.now()),
-                new Weather(region2, "2", -4, LocalDateTime.now())
+                new Weather(region1, "1", 1.0, LocalDateTime.now()),
+                new Weather(region2, "2", -2.0, LocalDateTime.now()),
+                new Weather(region2, "2", -4.0, LocalDateTime.now())
         );
         when(weatherRepository.findAll()).thenReturn(weather);
 
-        List<UUID> actual = service.getRegionsWithTemperatureAbove(0);
+        List<UUID> actual = service.getRegionsWithTemperatureAbove(0.0);
 
         List<UUID> expected = List.of(region1);
         assertEquals(expected, actual);
@@ -76,11 +76,11 @@ public class WeatherServiceImplTest {
         );
         when(weatherRepository.findAll()).thenReturn(weather);
 
-        Map<UUID, List<Integer>> actual = service.getMapWithId();
+        Map<UUID, List<Double>> actual = service.getMapWithId();
 
-        Map<UUID, List<Integer>> expected = Map.of(
-                region1, List.of(1),
-                region2, List.of(-2, -4)
+        Map<UUID, List<Double>> expected = Map.of(
+                region1, List.of(1.0),
+                region2, List.of(-2.0, -4.0)
         );
         assertEquals(expected, actual);
     }
@@ -89,17 +89,17 @@ public class WeatherServiceImplTest {
     void getMapWithTemperature_isCorrect() {
         UUID region1 = UUID.randomUUID();
         UUID region2 = UUID.randomUUID();
-        Weather weather1 = new Weather(region1, "1", 1, LocalDateTime.now());
-        Weather weather2 = new Weather(region2, "2", 1, LocalDateTime.now());
-        Weather weather3 = new Weather(region2, "2", -1, LocalDateTime.now());
+        Weather weather1 = new Weather(region1, "1", 1.0, LocalDateTime.now());
+        Weather weather2 = new Weather(region2, "2", 1.0, LocalDateTime.now());
+        Weather weather3 = new Weather(region2, "2", -1.0, LocalDateTime.now());
         List<Weather> weather = List.of(weather1, weather2, weather3);
         when(weatherRepository.findAll()).thenReturn(weather);
 
-        Map<Integer, List<Weather>> actual = service.getMapWithTemperature();
+        Map<Double, List<Weather>> actual = service.getMapWithTemperature();
 
-        Map<Integer, List<Weather>> expected = Map.of(
-                1, List.of(weather1, weather2),
-                -1, List.of(weather3)
+        Map<Double, List<Weather>> expected = Map.of(
+                1.0, List.of(weather1, weather2),
+                -1.0, List.of(weather3)
         );
         assertEquals(expected, actual);
     }
